test(fastapi): consolidate filter isolation regressions
Keep provider regressions in the existing provider test module so related behavior is reviewed and maintained together. Exercise real requests as well as OpenAPI: distinct configurations retain their own search and date filters, and custom pagination defaults agree with runtime limits and offsets. Equal configurations still reuse their dependency.

All 200 FastAPI extension tests pass, along with full mypy and Pyright. The production isolation fix preserves normal defaults and adds only one small aggregate call per request.
